feat: Allow username to be specified for connections
For both Redis and Sentinel connections, there was no way to specify a
password along with the username. All connections to redis could only go
through the default account. In some setup configurations, the default
account may be disabled in favor of using individual usernames and
limiting access to what each user may have.

In this commit, the username parameter (in the config), along with CLI
parameters `--redis-username` and `--sentinel-username` have been added
to support this use case. The docker entrypoint.sh has also been
modified to accomodate the new parameters. The hosts environment
variable has been left unchanged due to it's complexity and strict
format. I felt that it would be a breaking change to modify it in a way
that makes sense. It has been left as is for backwards compatibility.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>